# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: Trainers Hive — training marketplace API
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /session/switch:
    post:
      operationId: switchUser
      tags: [session]
      summary: Switch the active demo user (vendor / trainer / admin)
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/SwitchUserBody"
      responses:
        "200":
          description: Switched
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/CurrentUser"
...
  /hire-inquiries:
    post:
      operationId: createHireInquiry
      tags: [hire]
      summary: Submit a managed-service inquiry (Hire Us)
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateHireInquiryBody"
      responses:
        "201":
          description: Inquiry created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/HireInquiry"
    get:
      operationId: listHireInquiries
      tags: [hire]
      summary: List all inquiries (admin only)
      responses:
        "200":
          description: List of inquiries
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/HireInquiry"
...
  /hire-inquiries/{id}:
    get:
      operationId: getHireInquiry
      tags: [hire]
      summary: Get a single inquiry (admin or owner)
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: string
      responses:
        "200":
          description: Inquiry
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/HireInquiry"
        "404":
          description: Not found
  /hire-inquiries/{id}/status:
    patch:
      operationId: updateHireInquiryStatus
      tags: [hire]
      summary: Update inquiry status (admin only)
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UpdateHireInquiryStatusBody"
      responses:
        "200":
          description: Updated inquiry
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/HireInquiry"
  /hire-inquiries/{id}/messages:
    get:
      operationId: listHireInquiryMessages
      tags: [hire]
      summary: List messages on an inquiry (admin or owner)
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: string
      responses:
        "200":
          description: Messages
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/HireInquiryMessage"
    post:
      operationId: sendHireInquiryMessage
      tags: [hire]
      summary: Send a message on an inquiry (admin or owner)
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/SendHireInquiryMessageBody"
      responses:
        "201":
          description: Created message
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/HireInquiryMessage"
...
components:
  schemas:
...
    UserRole:
      type: string
      enum: [vendor, trainer, admin]
    CurrentUser:
      type: object
      required: [id, role, name, email]
      properties:
        id:
          type: string
        role:
          $ref: "#/components/schemas/UserRole"
        name:
          type: string
        email:
          type: string
        avatarUrl:
          type: string
        vendorId:
          type: string
        trainerId:
          type: string
    SwitchUserBody:
      type: object
      required: [role]
      properties:
        role:
          $ref: "#/components/schemas/UserRole"
        name:
          type: string
        email:
          type: string
          format: email
        orgName:
          type: string
        orgType:
          type: string
...
    HireInquiry:
      type: object
      required: [id, companyName, contactName, email, trainingNeed, status, createdAt]
      properties:
        id:
          type: string
        userId:
          type: string
        companyName:
          type: string
        contactName:
          type: string
        email:
          type: string
        phone:
          type: string
        trainingNeed:
          type: string
        budget:
          type: string
        timeline:
          type: string
        headcount:
          type: string
        location:
          type: string
        status:
          type: string
          enum: [new, contacted, in_progress, resolved, closed]
        createdAt:
          type: string
          format: date-time
    CreateHireInquiryBody:
      type: object
      required: [companyName, contactName, email, trainingNeed]
      properties:
        companyName:
          type: string
        contactName:
          type: string
        email:
          type: string
        phone:
          type: string
        trainingNeed:
          type: string
        budget:
          type: string
        timeline:
          type: string
        headcount:
          type: string
        location:
          type: string
    UpdateHireInquiryStatusBody:
      type: object
      required: [status]
      properties:
        status:
          type: string
          enum: [new, contacted, in_progress, resolved, closed]
    HireInquiryMessage:
      type: object
      required: [id, inquiryId, senderUserId, body, createdAt]
      properties:
        id:
          type: string
        inquiryId:
          type: string
        senderUserId:
          type: string
        body:
          type: string
        createdAt:
          type: string
          format: date-time
    SendHireInquiryMessageBody:
      type: object
      required: [body]
      properties:
        body:
          type: string
...